requirements.txt
Django>=3.0,<4.0.0.0
django-bootstrap4
django-cleanup
django-jsonify
django-dbbackup
django-debug-toolbar
django-phone-field
django-mptt
psycopg2>=2.7.0.0,<3.0.0.0
whitenoise
pillow
